LST and AT comparison maps
The study area for this map is the state of Arizona and the analysis level is Arizona's Primary Care Areas. The set of maps compares the LST obtained using Landsat imagery for the summer months between 2016-2020 and the Air Temp (AT) during the same time. The accompanying Tree Canopy Cover raster highlights the relationship between higher tree canopy cover and lower LS and Air temperatures.